Remove VOC algorithm
This just measures how close current values are to the last day's worth
of readings. This means that readings gradually drop during a long
print. This _looks_ like VOCs are decreasing, but it just means the
historical average is increasing. It also means that two sensors' VOC
readings cannot bear compared apples-to-apples. This also makes the
sensor values effectively useless for the first 24-hours after
initialization.

I think it makes more sense to report the current raw reading as a
percentage of a "normal" reading.  This prevents problems with
hysteresis and keeps sensors' readings in comparable domains.

In the future, I'd like to create a calibration command, that samples
sensors' raw values over a period of time and writes an appropriate
baseline to the config file.